order model created

+++ src/app/modules/product/product.services.ts
@@ -1,9 +1,10 @@
 import QueryBuilder from "../../builder/QueryBuilder";
+import AppError from "../../errors/AppError";
 import { IProduct } from "./product.interface";
 import { Product } from "./product.model";
 
-const createProductIntoDB = (payload: IProduct) => {
-  const result = Product.create(payload);
+const createProductIntoDB = async (payload: IProduct) => {
+  const result = await Product.create(payload);
   return result;
 };
 
@@ -19,7 +20,16 @@ const getAllProductsFromDB = async (query: Record<string, unknown>) => {
   return result;
 };
 
-const updateProductIntoDB = (productId: string, payload: Partial<IProduct>) => {
+const updateProductIntoDB = async (
+  productId: string,
+  payload: Partial<IProduct>
+) => {
+  const product = await Product.findById(productId);
+
+  if (!product) {
+    throw new AppError(404, "Product not found");
+  }
+
   const result = Product.findByIdAndUpdate(productId, payload, {
     new: true,
     runValidators: true,
@@ -27,7 +37,12 @@ const updateProductIntoDB = (productId: string, payload: Partial<IProduct>) => {
   return result;
 };
 
-const deleteProductFromDB = (productId: string) => {
+const deleteProductFromDB = async (productId: string) => {
+  const product = await Product.findById(productId);
+
+  if (!product) {
+    throw new AppError(404, "Product not found");
+  }
   const result = Product.findByIdAndDelete(productId);
   return result;
 };
